+++ terraform/modules/cloudfront/main.tf
@@ -10,6 +10,12 @@ resource "aws_cloudfront_distribution" "main" {
     target_origin_id       = "s3-origin"
     allowed_methods        = ["GET", "HEAD"]
     cached_methods         = ["GET", "HEAD"]
+    forwarded_values {
+      query_string = false
+      cookies {
+        forward = "none"
+      }
+    }
   }
 
   restrictions {